refactor polls into separate files for better organization
Split monolithic 918-line polls.go into 4 focused files:
- poll_types.go: data structures and state management
- poll_logic.go: core voting business logic and IRV algorithm
- poll_handlers.go: Discord interaction handlers
- poll_render.go: templates, rendering, and utilities

This improves code maintainability and readability by separating concerns.
